+++ Sources/MacApps/MacAppsApp.swift
@@ -5,8 +5,8 @@ struct MacAppsApp: App {
     @StateObject private var appState = AppState()
     @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
 
-    static let version = "0.3.3.0"
-    static let buildDate = "2024-12-24"
+    static let version = "0.3.5.0"
+    static let buildDate = "2025-12-25"
 
     var body: some Scene {
         WindowGroup {